+++ src/test/java/spofo/small/portfolio/service/PortfolioServiceTest.java
@@ -21,6 +21,7 @@ import org.junit.jupiter.api.Test;
 import spofo.global.domain.exception.PortfolioNotFound;
 import spofo.mock.FakePortfolioRepository;
 import spofo.mock.FakeStockServerService;
+import spofo.portfolio.controller.port.PortfolioService;
 import spofo.portfolio.domain.Portfolio;
 import spofo.portfolio.domain.PortfolioCreate;
 import spofo.portfolio.domain.PortfolioStatistic;
@@ -29,7 +30,6 @@ import spofo.portfolio.domain.PortfoliosStatistic;
 import spofo.portfolio.domain.enums.Currency;
 import spofo.portfolio.domain.enums.IncludeType;
 import spofo.portfolio.domain.enums.PortfolioType;
-import spofo.portfolio.service.PortfolioService;
 import spofo.portfolio.service.PortfolioServiceImpl;
 import spofo.stock.domain.Stock;
 import spofo.stockhave.domain.StockHave;
