+++ src/test/java/spofo/support/service/ServiceTestSupport.java
@@ -1,12 +1,12 @@
 package spofo.support.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import spofo.portfolio.service.PortfolioService;
+import spofo.portfolio.controller.port.PortfolioService;
 import spofo.portfolio.service.port.PortfolioRepository;
 import spofo.stock.service.StockServerService;
-import spofo.stockhave.service.StockHaveService;
+import spofo.stockhave.controller.port.StockHaveService;
 import spofo.support.annotation.CustomServiceTest;
-import spofo.tradelog.service.TradeLogService;
+import spofo.tradelog.controller.port.TradeLogService;
 
 @CustomServiceTest
 public abstract class ServiceTestSupport {
